+++ NEWS.md
@@ -1,4 +1,10 @@
-# rgl 1.0.3
+# rgl 1.0.4
+
+## Major changes
+
+* A new function `hover3d()` has been added to display
+"hover hints":  labels next to points when the mouse 
+passes near them.
 
 ## Bug fixes